@@ -1,13 +1,13 @@
 package com.cube.nanotimer.util.helper;
 
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import junit.framework.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.util.List;
 
-import static android.support.test.InstrumentationRegistry.getContext;
+import static androidx.test.platform.app.InstrumentationRegistry.getContext;
 
 @RunWith(AndroidJUnit4.class)
 public class FileUtilsTest {
