@@ -2,7 +2,7 @@ package com.cube.nanotimer.gui.widget;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import com.cube.nanotimer.util.helper.Utils;
 
 public class NanoTimerDialogFragment extends DialogFragment {
